# api/internal/handlers/swagger.yaml
openapi: 3.0.3
info:
  title: StreamSpace API
  description: |
    StreamSpace is a platform-agnostic container streaming platform that delivers GUI applications to web browsers.

    ## Architecture
    - **Control Plane**: Centralized API with WebSocket Hub for agent coordination
    - **Agents**: Lightweight executors on target platforms (Kubernetes, Docker)
    - **VNC Proxy**: Secure, firewall-friendly VNC streaming through WebSocket

    ## Authentication
    - **Users**: JWT tokens (24-hour expiration) via `/api/v1/auth/login`
    - **Agents**: API keys via `/api/v1/agents/register`
    - **Webhooks**: HMAC signatures

    ## Rate Limiting
    Default: 60 requests/minute per IP (configurable via `RATE_LIMIT_REQUESTS_PER_MINUTE`)

    ## Common Response Formats
    All error responses follow:
    ```json
    {
      "error": "ErrorCode",
      "message": "Detailed error message"
    }
    ```
  version: 2.0.0-beta
  contact:
    name: StreamSpace Team
    url: https://github.com/streamspace-dev/streamspace
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.streamspace.example.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for agent authentication
...
paths:
...
  /api/v1/plugins:
    get:
      summary: List installed plugins
      operationId: listPlugins
      tags:
        - plugins
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Installed plugins

    post:
      summary: Install plugin
      operationId: installPlugin
      tags:
        - plugins
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                  description: Plugin source URL or marketplace ID
      responses:
        '201':
          description: Plugin installed
  /api/v1/plugins/{id}:
    get:
      summary: Get plugin details
      operationId: getPlugin
      tags:
        - plugins
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plugin details

    delete:
      summary: Uninstall plugin
      operationId: uninstallPlugin
      tags:
        - plugins
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Plugin uninstalled
  /api/v1/plugins/{id}/enable:
    post:
      summary: Enable plugin
      operationId: enablePlugin
      tags:
        - plugins
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plugin enabled
  /api/v1/plugins/{id}/disable:
    post:
      summary: Disable plugin
      operationId: disablePlugin
      tags:
        - plugins
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plugin disabled
...